Resolved conflicts + graphql organization
Fixed
  - ManageInvites updated imports for mui-pro
  - Home page had 2 Slider imports

Changed
  - Abstracted gql and put them into graphql folder respectively
  - Linted along the way

Note
  - This will probably cause more conflicts
