+++ src/hhsbviews/HomePage.js
@@ -1,9 +1,8 @@
 import React from 'react'
 
-import Slider from '@material-ui/core/Slider';
+import Slider from '@material-ui/core/Slider'
 import { getThemeProps } from '@material-ui/styles'
 import SearchIcon from '@material-ui/icons/Search'
-import Slider from '@material-ui/core/Slider'
 import CalendarTodayTwoToneIcon from '@material-ui/icons/CalendarTodayTwoTone'
 import Box from '@material-ui/core/Box'
 
@@ -23,23 +22,18 @@ import Filter from 'hhsbAssets/Filter.svg'
 import Emoji from 'hhsbAssets/FollowingEmoji.svg'
 
 import SnackbarContent from "mui-pro/Snackbar/SnackbarContent.js"
-import Pagination from "material-ui-flat-pagination";
 
 import CustomizedInputBase from 'hhsbComponents/searchBar.js'
 import AlertList from  'hhsbComponents/AlertList.js'
-import CustomizedInputBase from 'hhsbComponents/searchBar.js'
-import AlertList from  'hhsbComponents/AlertList.js'
-import Content from "../hhsbComponents/ContentList.js"
 
-import Calendar from 'hhsbAssets/Calendar.svg'
-import Filter from 'hhsbAssets/Filter.svg'
-import Emoji from 'hhsbAssets/FollowingEmoji.svg'
 
 export default function HomePage() {
   let NotificationData=
    [{AlertTitle:'Quoted',color:'#00CAE3', AlertBody:'There’s no time like the present, so present your time as if it only one in the world”',time:'Today @ 10pm',points:'+100W'},{AlertTitle:'Upvoted',color:'#55B559', AlertBody:'',time:'',points:''},
      {AlertTitle:'Commented',color:'#FF9E0F', AlertBody:'',time:'',points:''},{AlertTitle:'Downvoted',color:'#FF1100', AlertBody:'',time:'',points:''},
-     {AlertTitle:'Submitted',color:'#000000', AlertBody:'',time:'',points:''},{AlertTitle:'Heared',color:'#E91E63', AlertBody:'',time:'',points:''}]
+     {AlertTitle:'Submitted',color:'#000000', AlertBody:'',time:'',points:''},{AlertTitle:'Heared',color:'#E91E63', AlertBody:'',time:'',points:''},
+   ]
+
   const handleClick=(x)=>{
     // console.log(x)
   }
@@ -70,23 +64,14 @@ export default function HomePage() {
           <GridContainer
             alignItems="center"  direction="row"
             justify="space-between"  style={{backgroundColor:"#FF7A00",boxShadow:"0 6px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19)",width:"75%",wrap:"nowrap"}}>  
-            <h3 style={{color:"white",font:"League Spartan",fontWeight: "bold",paddingLeft:"20px",paddingBottom:"5px"}}>Activity Feed</h3>                   
-                                                
-               
+            <h3 style={{color:"white",font:"League Spartan",fontWeight: "bold",paddingLeft:"20px",paddingBottom:"5px"}}>Activity Feed</h3>                                                   
             <div style={{display:'flex',justifyContent:'flex-end',flexDirection:"row"}}>
               <CustomizedInputBase></CustomizedInputBase>
               <img src={Calendar} style={{display:"flex",maxHeight:"40px",paddingLeft:"15px"}}/>
               <img src={Filter} style={{display:"flex",maxHeight:"40px",paddingLeft:"15px"}}/>
               <img src={Emoji} style={{display:"flex",maxHeight:"40px",paddingLeft:"15px",paddingRight:"15px"}}/> 
             </div>
-                      
-                 
-                
-
-                 
-            
-          </GridContainer> 
-                
+          </GridContainer>    
         </GridContainer>
         <br></br>
         <br></br>
@@ -99,9 +84,5 @@ export default function HomePage() {
         onClick={(e, offset) => this.handleClick(offset)}
       />
     </Card>
-
-
   )
-
-
 }
